+++ test/workers/send_site_setup_emails_test.exs
@@ -1,20 +1,16 @@
 defmodule Plausible.Workers.SendSiteSetupEmailsTest do
   use Plausible.DataCase
   use Bamboo.Test
+  use Oban.Testing, repo: Plausible.Repo
   import Plausible.TestUtils
   alias Plausible.Workers.SendSiteSetupEmails
-
-  defp perform() do
-    SendSiteSetupEmails.new(%{}) |> Oban.insert!()
-    Oban.drain_queue(:site_setup_emails)
-  end
 
   describe "when user has not managed to set up the site" do
     test "does not send an email 47 hours after site creation" do
       user = insert(:user)
       insert(:site, members: [user], inserted_at: hours_ago(47))
 
-      perform()
+      perform_job(SendSiteSetupEmails, %{})
 
       assert_no_emails_delivered()
     end
@@ -23,7 +19,7 @@ defmodule Plausible.Workers.SendSiteSetupEmailsTest do
       user = insert(:user)
       insert(:site, members: [user], inserted_at: hours_ago(49))
 
-      perform()
+      perform_job(SendSiteSetupEmails, %{})
 
       assert_email_delivered_with(
         to: [{user.name, user.email}],
@@ -35,7 +31,7 @@ defmodule Plausible.Workers.SendSiteSetupEmailsTest do
       user = insert(:user)
       insert(:site, members: [user], inserted_at: hours_ago(73))
 
-      perform()
+      perform_job(SendSiteSetupEmails, %{})
 
       assert_no_emails_delivered()
     end
@@ -46,7 +42,7 @@ defmodule Plausible.Workers.SendSiteSetupEmailsTest do
       user = insert(:user)
       insert(:site, members: [user], domain: "test-site.com")
 
-      perform()
+      perform_job(SendSiteSetupEmails, %{})
 
       assert_email_delivered_with(
         to: [{user.name, user.email}],
@@ -58,7 +54,7 @@ defmodule Plausible.Workers.SendSiteSetupEmailsTest do
       user = insert(:user)
       site = insert(:site, members: [user], inserted_at: hours_ago(49))
 
-      perform()
+      perform_job(SendSiteSetupEmails, %{})
 
       assert_email_delivered_with(
         to: [{user.name, user.email}],
@@ -66,7 +62,7 @@ defmodule Plausible.Workers.SendSiteSetupEmailsTest do
       )
 
       create_pageviews([%{domain: site.domain}])
-      perform()
+      perform_job(SendSiteSetupEmails, %{})
 
       assert_email_delivered_with(
         to: [{user.name, user.email}],
@@ -79,7 +75,7 @@ defmodule Plausible.Workers.SendSiteSetupEmailsTest do
     test "does not send an email before 48h have passed" do
       insert(:user, inserted_at: hours_ago(47))
 
-      perform()
+      perform_job(SendSiteSetupEmails, %{})
 
       assert_no_emails_delivered()
     end
@@ -87,7 +83,7 @@ defmodule Plausible.Workers.SendSiteSetupEmailsTest do
     test "sends the create site email after 48h" do
       user = insert(:user, inserted_at: hours_ago(49))
 
-      perform()
+      perform_job(SendSiteSetupEmails, %{})
 
       assert_email_delivered_with(
         to: [{user.name, user.email}],